controller/requirements.txt
PySerial
paramiko